Image_Search/Image_Search/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
Cython==0.29.28
gensim==4.1.2
h11==0.13.0
icrawler==0.6.6
idna==3.3
joblib==1.1.0
lxml==4.8.0
nltk==3.7
numpy==1.22.3
opencv-contrib-python==4.5.5.64
outcome==1.1.0
Pillow==9.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
threadpoolctl==3.1.0
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wikipedia==1.4.0
wsproto==1.1.0
